Rename pvalue to alpha, update documentation, and incorporate interpretation function
Alpha is a more traditional name for the threshold value. The value output from tests is usually called the pvalue.

The documentation was updated to reflect that ADF is a unit-root test, which may suggest nonstaionarity. Changed accept the null to "fail to reject" the null hypothesis. That's the most we can do with a pvalue comparison.

Incorporated the interpretation function and named it interpreted_stationarity_tests. This outputs a meaningfully consistent interpretation for the results from a list of stationarity tests. It also includes detailed documentation on how the interpretations are reached. This is useful because ADF and KPSS are technically testing for different forms of stationarity and keeping track of their combined meaning is difficult. The output also provides a suggested solution to reach stationarity.