hotfix for stationarity unit test
From an email on this change:

The ADF test on the deterministic linear trend should be True, not False. It makes sense now why it returns adf:True and kpss:False. The deterministic trend is sort of the definition of “trend stationarity,” which is that the trend is deterministic. It’s in that case that we should get adf:True and kpss:False. What I don’t understand is why the stochastic linear trend does not return trend stationarity, it just says not stationary. In either case, these test are not perfect and the parameters they accept are more about asking specific stationarity questions rather than tuning them. I think we should code into the unit tests what the tests return right now. If people use these they will need to understand their interpretability and so we should just ensure their consistency.